+++ examples/advanced/main.tf
@@ -1,6 +1,7 @@
 locals {
   naming_prefix = "test-github-runner"
   vpc_id        = "vpc-0ffaabbcc1122"
+  vpc_cidr      = "10.0.0.0/16"
 }
 
 # Create a custom security-group to allow SSH to all EC2 instances
@@ -22,18 +23,11 @@ resource "aws_security_group" "this" {
   #checkov:skip=CKV_AWS_382:Egress to GitHub Actions is required for the runner to work.
 }
 
-data "http" "myip" {
-  url = "http://ipv4.icanhazip.com"
-}
-
-resource "aws_security_group_rule" "ssh_ingress" {
-  description       = "Allow SSH ingress to EC2 instance"
-  type              = "ingress"
-  from_port         = 22
-  to_port           = 22
-  protocol          = "tcp"
-  cidr_blocks       = ["${chomp(data.http.myip.body)}/32"]
-  security_group_id = aws_security_group.this.id
+# Create a baseline CodeBuild credential that all GitHub projects will use by default
+resource "aws_codebuild_source_credential" "github" {
+  auth_type   = "SECRETS_MANAGER"
+  server_type = "GITHUB"
+  token       = "arn:aws:secretsmanager:region:account-id:secret:name"
 }
 
 module "github_runner" {
@@ -41,8 +35,7 @@ module "github_runner" {
 
   # Required parameters
   ############################
-  source_location                            = "https://github.com/my-org/my-repo.git"
-  github_personal_access_token_ssm_parameter = "example"
+  source_location = "https://github.com/my-org/my-repo.git"
 
   # Naming for all created resources
   name = "github-runner-codebuild-test"
@@ -60,6 +53,44 @@ module "github_runner" {
 
   create_ecr_repository = true
 
+  # Override the baseline CodeBuild credential
+  source_auth = {
+    type     = "SECRETS_MANAGER"
+    resource = "arn:aws:secretsmanager:af-south-1:123456789012:secret:my-github-oauth-token-secret-nwYBWW"
+  }
+
   security_group_ids         = [aws_security_group.this.id]
   cloudwatch_logs_group_name = "/some/log/group"
+}
+
+# Example: Using the default security group with custom ingress rules for Packer
+module "github_runner_with_packer" {
+  source = "../../"
+
+  # Required parameters
+  source_location = "https://github.com/my-org/my-repo.git"
+  name            = "github-runner-packer"
+
+  # VPC configuration
+  vpc_id     = local.vpc_id
+  subnet_ids = ["subnet-0123", "subnet-0456"]
+
+  # Custom ingress rules added to the default security group
+  # This is useful when running Packer which requires ephemeral ports for WinRM/SSH
+  ingress_with_cidr_blocks = [
+    {
+      from_port   = 1024
+      to_port     = 65535
+      protocol    = "tcp"
+      description = "Ephemeral ports required for Packer WinRM/SSH communication"
+      cidr_blocks = [local.vpc_cidr]
+    },
+    {
+      from_port   = 5985
+      to_port     = 5986
+      protocol    = "tcp"
+      description = "WinRM ports for Packer"
+      cidr_blocks = [local.vpc_cidr]
+    }
+  ]
 }